Task 11 Write a program that finds and prints the largest prime factor of the number 612852475143, followed by a new line.

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -5,7 +5,7 @@
 */
 int main(void)
 {
-long long num, max, i;
+long num, max, i;
 num = 612852475143;
 max = 1;
 while (num % 2 == 0)
@@ -21,6 +21,6 @@ max = i;
 num /= i;
 }
 }
-printf("%lld\n", max);
+printf("%lu\n", max);
 return (0);
 }
